config.php있으면 설치 못하게 하게 고침

+++ install.php
@@ -10,6 +10,9 @@ include_once("soojung.php");
 <body>
 
 <?php
+if (file_exists('config.php')) {
+  die("already installed");
+}
 if (!is_writable(".")) {
   echo "<font color=\"red\">WARNING: This directory is not writeable</font><br />";
 }
